Consolidate Spring Boot WebMVC smoke tests
